requirements.txt
anyio==3.7.1
APScheduler==3.10.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
copier==8.1.0
cryptography==41.0.2
decorator==5.1.1
dunamai==1.18.0
exceptiongroup==1.1.2
flet==0.8.4
flet-core==0.8.4
flet-runtime==0.8.4
funcy==2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
plumbum==1.8.2
prompt-toolkit==3.0.39
protobuf==4.21.12
PYBRDOC==1.1.5
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
pypng==0.20220715.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==1.10.0
repath==0.9.0
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
